test: isolate Include environment coverage
